feat: source chain registry relationships + executor access control (#359)
* feat: implement source-chain relationship management in CrossChainRegistry

- Introduced a new `EXECUTOR` relationship type to manage authorized executors on the same chain as the `BridgeRouter`.
- Added functions for registering and removing executors, enhancing the security and clarity of cross-chain calls.
- Updated the `registerSourceChainRelationship` function to allow same-chain relationships, streamlining the relationship management process.
- Created a new documentation file outlining the plan for making `BridgeRouter` the single choke-point for cross-chain calls, detailing the registry and executor management.
- Enhanced unit tests to validate the new functionalities and ensure proper authorization checks for executors.

* feat: enhance CrossChainRegistry and BridgeRouter integration

- Added a new `EXECUTOR` constant in `CrossChainRegistry` for managing executor relationships.
- Updated `IBridgeRouter` to include a more descriptive error for unauthorized calls, changing `OnlyAuthorized` to `OnlyAuthorizedExecutor`.
- Integrated `ICrossChainRegistry` into `BridgeRouter`, allowing it to reference the executor relationship for authorization checks.
- Modified the `BridgeRouter` constructor to accept a registry address, ensuring proper initialization.
- Updated tests to validate the new registry integration and executor authorization logic, enhancing overall functionality and security.

* feat: enhance CrossChainRegistry and BridgeRouter functionality

- Introduced logic to differentiate between same-chain and inter-chain relationships in CrossChainRegistry, improving contract clarity and relationship management.
- Updated the target contract registration process to ensure uniqueness only for inter-chain relationships, enhancing security.
- Added a new `deliver` function in IBridgeRouter for handling inbound asset transfers and messages, streamlining cross-chain operations.
- Implemented comprehensive tests for the new `deliver` function, validating asset transfers and message deliveries, ensuring robust functionality across adapters.

* chore: remove source_chain_access_control.md documentation file

- Deleted the `source_chain_access_control.md` file, which outlined the plan for making `BridgeRouter` the single choke-point for cross-chain calls, as it is no longer needed.
- Updated the openzeppelin-contracts-upgradeable and pendle-core-v2-public subproject commits to indicate dirty states, reflecting local modifications.

* refactor: streamline CrossChainRegistry and ICrossChainRegistry documentation

- Replaced detailed function comments with `@inheritdoc` tags in CrossChainRegistry to enhance code clarity and maintainability.
- Updated ICrossChainRegistry interface to include function documentation for registering and managing executors, improving overall documentation consistency.

* refactor: simplify comments in CrossChainRegistry for improved clarity

- Removed redundant comment sections in CrossChainRegistry to streamline the code and enhance readability.
- Retained essential comments to maintain context while reducing visual clutter, improving overall maintainability.

* refactor: introduce BridgeRouterSetup for streamlined test setup

- Created a new BridgeRouterSetup contract to encapsulate common setup logic for BridgeRouter tests, enhancing code reusability and maintainability.
- Updated existing test contracts (BridgeRouterAdaptersTest, BridgeRouterAdminTest, BridgeRouterDeliverTest, BridgeRouterReadStateTest, and BridgeRouterTransferTest) to inherit from BridgeRouterSetup, simplifying their setup processes.
- Removed redundant setup code from individual test contracts, improving overall clarity and reducing duplication.

* refactor: update FleetProxy and MockBridgeRouter for improved parameter clarity

- Renamed parameters in FleetProxy from `hubChainId` to `_hubChainId` for consistency and clarity.
- Refactored the `withdrawAndTransfer` function to utilize a structured `ExecuteTransferParams` for asset transfers, enhancing readability and maintainability.
- Updated the MockBridgeRouter contract to remove the unused `payload` parameter, streamlining the function signature.
- Enhanced comments throughout the code to improve understanding of the logic and flow.

* refactor: clean up imports and formatting across multiple contracts

- Removed unused imports and organized import statements for clarity in various contracts, including BaseBridgeAdapter, LayerZeroAdapter, and StargateAdapter.
- Standardized formatting by adjusting whitespace and line breaks to enhance code readability.
- Updated the build-unit-test workflow by removing the lint step to streamline the CI process.
- Marked subproject commits for openzeppelin-contracts-upgradeable and pendle-core-v2-public as dirty to reflect local modifications.

* chore: format all

* chore: ignore formatting in /lib'

* refactor: standardize relationship registration in CrossChainRegistry

- Renamed `registerCrossChainRelationship` to `registerRelationship` for consistency across the codebase.
- Updated references to `ADAPTER_PEER` to `PEER` in various contracts and tests to align with the new naming convention.
- Introduced `isExecutor` function in `ICrossChainConfigManaged` to check executor status, enhancing contract functionality.
- Improved clarity and maintainability by refactoring related comments and documentation throughout the affected files.

* refactor: clean up imports and linting across contracts and tests

- Updated import statements in StargateAdapter, CrossChainRegistry, and various test files to remove unused imports, enhancing code clarity.
- Adjusted linting directives for unused imports and mixed-case variables to improve consistency and readability.
- Renamed variables in tests to follow consistent naming conventions, ensuring better maintainability.
- Streamlined the codebase by removing unnecessary comments and improving overall organization.

+++ packages/chain-bridge/src/adapters/LayerZeroAdapter.sol
@@ -1,24 +1,26 @@
 // SPDX-License-Identifier: BUSL-1.1
 pragma solidity 0.8.28;
 
-import {IERC20} from "@openzeppelin/contracts/token/ERC20/IERC20.sol";
-import {Ownable} from "@openzeppelin/contracts/access/Ownable.sol";
+import {LayerZeroOptionsHelper} from "../helpers/LayerZeroOptionsHelper.sol";
 import {IBridgeAdapter} from "../interfaces/IBridgeAdapter.sol";
 import {IBridgeRouter} from "../interfaces/IBridgeRouter.sol";
 /// forge-lint: disable-next-item(unused-import)
 import {ISendAdapter} from "../interfaces/ISendAdapter.sol";
 import {BridgeTypes} from "../libraries/BridgeTypes.sol";
-import {Origin} from "@layerzerolabs/oapp-evm/contracts/oapp/OApp.sol";
-import {LayerZeroOptionsHelper} from "../helpers/LayerZeroOptionsHelper.sol";
-import {SafeERC20} from "@openzeppelin/contracts/token/ERC20/utils/SafeERC20.sol";
-import {OAppRead} from "@layerzerolabs/oapp-evm/contracts/oapp/OAppRead.sol";
-import {ReadCodecV1, EVMCallRequestV1} from "@layerzerolabs/oapp-evm/contracts/oapp/libs/ReadCodecV1.sol";
-import {MessagingFee as EndpointFee, MessagingReceipt} from "@layerzerolabs/lz-evm-protocol-v2/contracts/interfaces/ILayerZeroEndpointV2.sol";
 import {ICrossChainMessageReceiver} from "../interfaces/ICrossChainMessageReceiver.sol";
-import {EnumerableSet} from "@openzeppelin/contracts/utils/structs/EnumerableSet.sol";
-import {SetConfigParam} from "@layerzerolabs/lz-evm-protocol-v2/contracts/interfaces/IMessageLibManager.sol";
-import {ReadLibConfig} from "@layerzerolabs/lz-evm-messagelib-v2/contracts/uln/readlib/ReadLibBase.sol";
 import {BaseBridgeAdapter} from "./BaseBridgeAdapter.sol";
+import {ReadLibConfig} from "@layerzerolabs/lz-evm-messagelib-v2/contracts/uln/readlib/ReadLibBase.sol";
+import {MessagingFee as EndpointFee, MessagingReceipt} from "@layerzerolabs/lz-evm-protocol-v2/contracts/interfaces/ILayerZeroEndpointV2.sol";
+import {SetConfigParam} from "@layerzerolabs/lz-evm-protocol-v2/contracts/interfaces/IMessageLibManager.sol";
+import {Origin} from "@layerzerolabs/oapp-evm/contracts/oapp/OApp.sol";
+import {OAppRead} from "@layerzerolabs/oapp-evm/contracts/oapp/OAppRead.sol";
+import {EVMCallRequestV1, ReadCodecV1} from "@layerzerolabs/oapp-evm/contracts/oapp/libs/ReadCodecV1.sol";
+import {Ownable} from "@openzeppelin/contracts/access/Ownable.sol";
+import {IERC20} from "@openzeppelin/contracts/token/ERC20/IERC20.sol";
+
+import {SafeERC20} from "@openzeppelin/contracts/token/ERC20/utils/SafeERC20.sol";
+
+import {EnumerableSet} from "@openzeppelin/contracts/utils/structs/EnumerableSet.sol";
 
 /**
  * @title LayerZeroAdapter
@@ -132,7 +134,7 @@ contract LayerZeroAdapter is OAppRead, IBridgeAdapter, BaseBridgeAdapter {
         if (_supportedChains.length != _lzEids.length) revert InvalidParams();
 
         // Setup chain ID mappings
-        for (uint i = 0; i < _supportedChains.length; i++) {
+        for (uint256 i = 0; i < _supportedChains.length; i++) {
             chainToLzEid[_supportedChains[i]] = _lzEids[i];
             lzEidToChain[_lzEids[i]] = _supportedChains[i];
             _supportedChainIds.add(_supportedChains[i]);
@@ -254,7 +256,7 @@ contract LayerZeroAdapter is OAppRead, IBridgeAdapter, BaseBridgeAdapter {
         if (executor == address(0)) revert InvalidParams();
 
         // Verify DVNs are sorted (required by LayerZero)
-        for (uint i = 1; i < readDVNs.length; i++) {
+        for (uint256 i = 1; i < readDVNs.length; i++) {
             if (readDVNs[i] <= readDVNs[i - 1]) revert InvalidParams(); // Must be sorted
         }
